requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
celery==5.3.4
redis==5.0.1
playwright==1.40.0
yt-dlp==2023.11.16
whisper==1.1.10
ffmpeg-python==0.2.0
requests==2.31.0
streamlit==1.28.2
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.0
pillow==10.1.0